pyproject.toml

[tool.poetry]
name = "telegram-pdf-bot"
version = "0.1.0"
description = "Telegram bot to convert HTML to PDF"
authors = ["You"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.8"
pdfkit = "^1.0.0"
pypdf2 = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
